+++ src/components/tables/cells/fieldSelection/types.ts
@@ -1,0 +1,11 @@
+import {
+    FieldSelectionType,
+    TranslatedConstraint,
+} from 'components/editor/Bindings/FieldSelection/types';
+
+export interface FieldActionsProps {
+    bindingUUID: string;
+    field: string;
+    constraint: TranslatedConstraint;
+    selectionType: FieldSelectionType | null;
+}
