+++ src/utils/workflow-utils.ts
@@ -3,7 +3,10 @@ import {
     DraftSpecsExtQuery_ByCatalogName,
     modifyDraftSpec,
 } from 'api/draftSpecs';
-import { ConstraintTypes } from 'components/editor/Bindings/FieldSelection/types';
+import {
+    ConstraintTypes,
+    FieldSelectionType,
+} from 'components/editor/Bindings/FieldSelection/types';
 import { ConnectorWithTagDetailQuery } from 'hooks/connectors/shared';
 import { DraftSpecQuery } from 'hooks/useDraftSpecs';
 import { isBoolean, isEmpty } from 'lodash';
@@ -398,6 +401,10 @@ export const isExcludeOnlyField = (
         constraintType === ConstraintTypes.UNSATISFIABLE
     );
 };
+
+export const isFieldSelectionType = (value: any): value is FieldSelectionType =>
+    typeof value === 'string' &&
+    (value === 'default' || value === 'exclude' || value === 'require');
 
 export interface ConnectorVersionEvaluationOptions {
     connectorId: string;
